[TASK] Include FPDM library from phar file
If TYPO3 instance has not been set up using composer, the FPDM library
is not available. which breaks functionality.

This change includes the library from a phar file, if the fully
qualified class name \FPDM does not exist.

It should be understood that this solution is a workaround only.

Conflicts:
	Classes/Pdf.php