Rename method createAction
The method createAction() is renamed to createActionBeforeRenderView()
and the class Pdf.php no longer extends from the Powermail class
FormController.

Previously, the createAction() was overridden with a different method
signature which results in exceptions with newer Powermail versions
due to strict.

The method name createActionBeforeRenderView() is now used to better
show that it is the signal slot method.